+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 ## 0.0.1
 
+Add basic function for http_miracle_mock
 
+## 0.0.2
+
+Change the warning code.
+
+## 0.0.3
+
+Pass static analysis